使用cstdint定义的数据类型重命名了相关的变量类型，此外还修正了命名规则，现在全都使用驼峰命名规则，放弃了c语言的类型转换方法现在一律都使用***_cast来转换类型

+++ Engine/Inc/Gear/Core/DX/ReadbackHeap.h
@@ -9,7 +9,7 @@ class ReadbackHeap :public Resource
 {
 public:
 
-	ReadbackHeap(const UINT size, const D3D12_HEAP_FLAGS flags = D3D12_HEAP_FLAG_NONE);
+	ReadbackHeap(const uint64_t size, const D3D12_HEAP_FLAGS flags = D3D12_HEAP_FLAG_NONE);
 
 	ReadbackHeap(ReadbackHeap&);
